+++ gpaw/output.py
@@ -96,6 +96,15 @@ class PAWTextOutput:
                   (os.path.dirname(ase.__file__), ase_version))
         self.text('numpy: %s (version %s)' %
                   (os.path.dirname(np.__file__), np.version.version))
+        try:
+            import scipy as sp
+            self.text('scipy: %s (version %s)' %
+                      (os.path.dirname(sp.__file__), sp.version.version))
+            # Explicitly deleting SciPy seems to remove garbage collection
+            # problem of unknown cause
+            del sp
+        except ImportError:
+            self.text('scipy: Not available')
         self.text('units: Angstrom and eV')
         self.text('cores:', self.world.size)
 
@@ -335,7 +344,7 @@ class PAWTextOutput:
             if eigerr == 0.0:
                 eigerr = ''
             else:
-                eigerr = '%-+5.2f' % (log(eigerr) / log(10))
+                eigerr = '%+.2f' % (log(eigerr) / log(10))
 
             denserr = self.density.mixer.get_charge_sloshing()
             if denserr is None or denserr == 0 or nvalence == 0:
@@ -354,7 +363,7 @@ class PAWTextOutput:
             else:
                 niterpoisson = str(self.hamiltonian.npoisson)
 
-            t('iter: %3d  %02d:%02d:%02d  %-5s  %-5s    %11.6f  %-5s  %-7s' %
+            t('iter: %3d  %02d:%02d:%02d %6s %6s    %11.6f  %-5s  %-7s' %
               (iter,
                T[3], T[4], T[5],
                eigerr,